Refactor simulation and popup message domains
