+++ app/ts/background/popupMessageHandlerRegistry.ts
@@ -1,0 +1,24 @@
+import type { EthereumClientService } from '../simulation/services/EthereumClientService.js'
+import type { TokenPriceService } from '../simulation/services/priceEstimator.js'
+import type { ResetSimulationServices } from '../simulation/serviceLifecycle.js'
+import type { PopupMessage, Settings } from '../types/interceptor-messages.js'
+import type { PopupReplyOption } from '../types/interceptor-reply-messages.js'
+import type { WebsiteTabConnections } from '../types/user-interface-types.js'
+import { createMethodHandlerFor } from '../utils/methodHandlers.js'
+import type { PublishRpcConnectionStatus } from './rpcSlowRequestTracking.js'
+
+export type PopupMessageDispatcherContext = {
+	websiteTabConnections: WebsiteTabConnections
+	ethereum: EthereumClientService
+	tokenPriceService: TokenPriceService
+	resetSimulationServices: ResetSimulationServices
+	settings: Settings
+	publishRpcConnectionStatus: PublishRpcConnectionStatus
+	simulationAbortController: AbortController
+	confirmTransactionAbortController: AbortController
+	resetSimulationState: () => Promise<void>
+}
+
+export type PopupMessageHandler = (context: PopupMessageDispatcherContext, request: PopupMessage) => Promise<PopupReplyOption | void>
+export type PopupMessageHandlerMap = Record<PopupMessage['method'], PopupMessageHandler>
+export const popupMessageHandler = createMethodHandlerFor<PopupMessage, PopupMessageDispatcherContext, Promise<PopupReplyOption | void>>()
