+++ app/ts/background/popupMessageHandlers/settings.ts
@@ -1,0 +1,53 @@
+import type { ImportSettings, ImportSettingsReply, SetRpcList, Settings } from '../../types/interceptor-messages.js'
+import { ExportedSettings } from '../../types/exportedSettingsTypes.js'
+import { serialize } from '../../types/wire-types.js'
+import { isJSON } from '../../utils/json.js'
+import { silenceChromeUnCaughtPromise } from '../../utils/requests.js'
+import type { ResetSimulationServices } from '../../simulation/serviceLifecycle.js'
+import { getPrimaryRpcForChain, getRpcList, setRpcList } from '../storageVariables.js'
+import { exportSettingsAndAddressBook, getMetamaskCompatibilityMode, getSettings, getUseTabsInsteadOfPopup, importSettingsAndAddressBook } from '../settings.js'
+import { sendPopupMessageToOpenWindows } from '../backgroundUtils.js'
+
+export async function settingsOpened() {
+	const useTabsInsteadOfPopupPromise = silenceChromeUnCaughtPromise(getUseTabsInsteadOfPopup())
+	const metamaskCompatibilityModePromise = silenceChromeUnCaughtPromise(getMetamaskCompatibilityMode())
+	const rpcEntriesPromise = silenceChromeUnCaughtPromise(getRpcList())
+	const settingsPromise = silenceChromeUnCaughtPromise(getSettings())
+
+	await sendPopupMessageToOpenWindows({
+		method: 'popup_requestSettingsReply' as const,
+		data: {
+			useTabsInsteadOfPopup: await useTabsInsteadOfPopupPromise,
+			metamaskCompatibilityMode: await metamaskCompatibilityModePromise,
+			rpcEntries: await rpcEntriesPromise,
+			activeRpcNetwork: (await settingsPromise).activeRpcNetwork
+		}
+	})
+}
+
+export async function importSettings(settingsData: ImportSettings): Promise<ImportSettingsReply> {
+	if (!isJSON(settingsData.data.fileContents)) {
+		return { method: 'popup_initiate_export_settings_reply', data: { success: false, errorMessage: 'Failed to read the file. It is not a valid JSON file.' } }
+	}
+	const parsed = ExportedSettings.safeParse(JSON.parse(settingsData.data.fileContents))
+	if (!parsed.success) {
+		return { method: 'popup_initiate_export_settings_reply', data: { success: false, errorMessage: 'Failed to read the file. It is not a valid interceptor settings file' } }
+	}
+	await importSettingsAndAddressBook(parsed.value)
+	return { method: 'popup_initiate_export_settings_reply', data: { success: true } }
+}
+
+export async function exportSettings() {
+	const exportedSettings = await exportSettingsAndAddressBook()
+	await sendPopupMessageToOpenWindows({
+		method: 'popup_initiate_export_settings',
+		data: { fileContents: JSON.stringify(serialize(ExportedSettings, exportedSettings), undefined, 4) }
+	})
+}
+
+export async function setNewRpcList(resetSimulationServices: ResetSimulationServices, request: SetRpcList, settings: Settings) {
+	await setRpcList(request.data)
+	await sendPopupMessageToOpenWindows({ method: 'popup_update_rpc_list', data: request.data })
+	const primary = await getPrimaryRpcForChain(settings.activeRpcNetwork.chainId)
+	if (primary !== undefined) resetSimulationServices(primary)
+}
