+++ app/ts/background/popupMessageHandlers/websiteAccess.ts
@@ -1,0 +1,106 @@
+import type { EthereumClientService } from '../../simulation/services/EthereumClientService.js'
+import type { TokenPriceService } from '../../simulation/services/priceEstimator.js'
+import type { ResetSimulationServices } from '../../simulation/serviceLifecycle.js'
+import type { AllowOrPreventAddressAccessForWebsite, BlockOrAllowExternalRequests, DisableInterceptor, RemoveWebsiteAccess, RemoveWebsiteAddressAccess, RetrieveWebsiteAccess } from '../../types/interceptor-messages.js'
+import type { EthereumAddress } from '../../types/wire-types.js'
+import type { Website } from '../../types/websiteAccessTypes.js'
+import { updateContentScriptInjectionStrategyManifestV2, updateContentScriptInjectionStrategyManifestV3 } from '../../utils/contentScriptsUpdating.js'
+import { getErrorMessage, reportUnexpectedError } from '../../utils/errors.js'
+import { checkAndThrowRuntimeLastError } from '../../utils/requests.js'
+import { modifyObject } from '../../utils/typescript.js'
+import { setInterceptorDisabledForWebsite, updateWebsiteApprovalAccesses } from '../accessManagement.js'
+import { sendPopupMessageToOpenWindows } from '../backgroundUtils.js'
+import { getLastKnownCurrentTabId } from '../currentTab.js'
+import { getSettings, updateWebsiteAccess } from '../settings.js'
+import type { WebsiteTabConnections } from '../../types/user-interface-types.js'
+import { getAddressMetadataForAccess } from '../windows/interceptorAccess.js'
+import { searchWebsiteAccess } from '../websiteAccessSearch.js'
+
+const isMissingTabReloadError = (error: unknown) => {
+	const message = getErrorMessage(error)
+	return message !== undefined && (message.startsWith('No tab with id') || message.includes('Invalid tab ID'))
+}
+
+export async function reloadConnectedTabs(websiteTabConnections: WebsiteTabConnections) {
+	const tabIdsToRefresh = Array.from(websiteTabConnections.keys())
+	const currentTabId = await getLastKnownCurrentTabId()
+	const withCurrentTabId = currentTabId === undefined ? tabIdsToRefresh : [...tabIdsToRefresh, currentTabId]
+	for (const tabId of new Set(withCurrentTabId)) {
+		try {
+			await browser.tabs.reload(tabId)
+			checkAndThrowRuntimeLastError()
+		} catch (error) {
+			if (isMissingTabReloadError(error)) continue
+			await reportUnexpectedError(error, { code: 'connected_tab_reload_failed' })
+		}
+	}
+}
+
+export const disableInterceptorForPage = async (websiteTabConnections: WebsiteTabConnections, website: Website, interceptorDisabled: boolean) => {
+	await setInterceptorDisabledForWebsite(website, interceptorDisabled)
+	if (browser.runtime.getManifest().manifest_version === 3) await updateContentScriptInjectionStrategyManifestV3()
+	else await updateContentScriptInjectionStrategyManifestV2()
+	await reloadConnectedTabs(websiteTabConnections)
+}
+
+export async function disableInterceptor(ethereum: EthereumClientService, tokenPriceService: TokenPriceService, resetSimulationServices: ResetSimulationServices, websiteTabConnections: WebsiteTabConnections, parsedRequest: DisableInterceptor) {
+	await disableInterceptorForPage(websiteTabConnections, parsedRequest.data.website, parsedRequest.data.interceptorDisabled)
+	await updateWebsiteApprovalAccesses(ethereum, tokenPriceService, resetSimulationServices, websiteTabConnections, await getSettings(), true)
+	await sendPopupMessageToOpenWindows({ method: 'popup_setDisableInterceptorReply' as const, data: parsedRequest.data })
+}
+
+export async function retrieveWebsiteAccess(parsedRequest: RetrieveWebsiteAccess) {
+	const settings = await getSettings()
+	const websiteAccess = searchWebsiteAccess(parsedRequest.data.query, settings.websiteAccess)
+	const addressAccessMetadata = await getAddressMetadataForAccess(websiteAccess)
+	await sendPopupMessageToOpenWindows({ method: 'popup_retrieveWebsiteAccessReply', data: { websiteAccess, addressAccessMetadata } })
+}
+
+const blockOrAllowWebsiteExternalRequests = async (websiteTabConnections: WebsiteTabConnections, website: Website, shouldBlock: boolean) => {
+	await updateWebsiteAccess((previousAccessList) => previousAccessList.map((access) => {
+		if (access.website.websiteOrigin !== website.websiteOrigin) return access
+		return modifyObject(access, { declarativeNetRequestBlockMode: shouldBlock ? 'block-all' : 'disabled' })
+	}))
+	await reloadConnectedTabs(websiteTabConnections)
+}
+
+export async function blockOrAllowExternalRequests(ethereum: EthereumClientService, tokenPriceService: TokenPriceService, resetSimulationServices: ResetSimulationServices, websiteTabConnections: WebsiteTabConnections, parsedRequest: BlockOrAllowExternalRequests) {
+	await blockOrAllowWebsiteExternalRequests(websiteTabConnections, parsedRequest.data.website, parsedRequest.data.shouldBlock)
+	await updateWebsiteApprovalAccesses(ethereum, tokenPriceService, resetSimulationServices, websiteTabConnections, await getSettings(), true)
+	await sendPopupMessageToOpenWindows({ method: 'popup_websiteAccess_changed' })
+}
+
+const removeAddressAccessByAddress = async (websiteOrigin: string, address: EthereumAddress) => {
+	await updateWebsiteAccess((previousAccessList) => previousAccessList.map((access) => {
+		if (access.website.websiteOrigin !== websiteOrigin || !access.addressAccess) return access
+		return modifyObject(access, { addressAccess: access.addressAccess.filter((addressAccess) => addressAccess.address !== address) })
+	}))
+}
+
+export async function removeWebsiteAddressAccess(ethereum: EthereumClientService, tokenPriceService: TokenPriceService, resetSimulationServices: ResetSimulationServices, websiteTabConnections: WebsiteTabConnections, parsedRequest: RemoveWebsiteAddressAccess) {
+	await removeAddressAccessByAddress(parsedRequest.data.websiteOrigin, parsedRequest.data.address)
+	await reloadConnectedTabs(websiteTabConnections)
+	await updateWebsiteApprovalAccesses(ethereum, tokenPriceService, resetSimulationServices, websiteTabConnections, await getSettings(), true)
+	await sendPopupMessageToOpenWindows({ method: 'popup_websiteAccess_changed' })
+}
+
+const setAddressAccessForWebsite = async (websiteOrigin: string, address: EthereumAddress, allowAccess: boolean) => {
+	await updateWebsiteAccess((previousAccessList) => previousAccessList.map((access) => {
+		if (access.website.websiteOrigin !== websiteOrigin || access.addressAccess === undefined) return access
+		const addressAccess = access.addressAccess.map((entry) => entry.address === address ? modifyObject(entry, { access: allowAccess }) : entry)
+		return modifyObject(access, { addressAccess })
+	}))
+}
+
+export async function allowOrPreventAddressAccessForWebsite(websiteTabConnections: WebsiteTabConnections, parsedRequest: AllowOrPreventAddressAccessForWebsite) {
+	const { website, address, allowAccess } = parsedRequest.data
+	await setAddressAccessForWebsite(website.websiteOrigin, address, allowAccess)
+	await reloadConnectedTabs(websiteTabConnections)
+	await sendPopupMessageToOpenWindows({ method: 'popup_websiteAccess_changed' })
+}
+
+export async function removeWebsiteAccess(ethereum: EthereumClientService, tokenPriceService: TokenPriceService, resetSimulationServices: ResetSimulationServices, websiteTabConnections: WebsiteTabConnections, parsedRequest: RemoveWebsiteAccess) {
+	await updateWebsiteAccess((previousAccess) => previousAccess.filter((access) => access.website.websiteOrigin !== parsedRequest.data.websiteOrigin))
+	await updateWebsiteApprovalAccesses(ethereum, tokenPriceService, resetSimulationServices, websiteTabConnections, await getSettings(), true)
+	await sendPopupMessageToOpenWindows({ method: 'popup_websiteAccess_changed' })
+}
