+++ app/ts/simulation/compoundGovernanceFaking.ts
@@ -6,7 +6,7 @@ import type { EthereumClientService } from './services/EthereumClientService.js'
 import { getCompoundGovernanceTimeLockMulticall } from '../utils/ethereumByteCodes.js'
 import * as funtypes from 'funtypes'
 import type { AddressBookEntry } from '../types/addressBookTypes.js'
-import { mockSignTransaction } from './services/SimulationModeEthereumClientService.js'
+import { mockSignTransaction } from './services/simulationTransactionSigning.js'
 import { DEFAULT_BLOCK_MANIPULATION } from '../config/defaults.js'
 import { decodeFunctionOutputLoose, decodeFunctionOutputObjectLoose, encodeFunctionCallLoose, hasFunctionLoose } from '../utils/abiRuntime.js'
