+++ app/ts/simulation/services/simulationPersonalSigning.ts
@@ -1,0 +1,31 @@
+import type { PersonalSignParams, SignMessageParams } from '../../types/jsonRpc-signing-types.js'
+import type { EthereumAddress } from '../../types/wire-types.js'
+import { bytes32String, stringToUint8Array } from '../../utils/bigint.js'
+import { hashMessage, hashTypedData, privateKeyToAccount } from '../../utils/ethereumPrimitives.js'
+import { assertNever } from '../../utils/typescript.js'
+
+const MOCK_PUBLIC_PRIVATE_KEY = 0x1n
+const MOCK_SIMULATION_PRIVATE_KEY = 0x2n
+const ADDRESS_FOR_PRIVATE_KEY_ONE = 0x7E5F4552091A69125d5DfCb7b8C2659029395Bdfn
+
+export const getMessageHashForPersonalSign = (params: PersonalSignParams) => hashMessage({ raw: stringToUint8Array(params.params[0]) })
+
+export const simulatePersonalSign = async (params: SignMessageParams, signingAddress: EthereumAddress) => {
+	const account = privateKeyToAccount(bytes32String(signingAddress === ADDRESS_FOR_PRIVATE_KEY_ONE ? MOCK_PUBLIC_PRIVATE_KEY : MOCK_SIMULATION_PRIVATE_KEY))
+	switch (params.method) {
+		case 'eth_signTypedData': throw new Error('No support for eth_signTypedData')
+		case 'eth_signTypedData_v1':
+		case 'eth_signTypedData_v2':
+		case 'eth_signTypedData_v3':
+		case 'eth_signTypedData_v4': {
+			const messageHash = hashTypedData(params.params[1])
+			const signature = await account.signTypedData(params.params[1])
+			return { signature, messageHash }
+		}
+		case 'personal_sign': return {
+			signature: await account.signMessage({ message: { raw: stringToUint8Array(params.params[0]) } }),
+			messageHash: getMessageHashForPersonalSign(params)
+		}
+		default: assertNever(params)
+	}
+}
